fix: :bug: Rename component plan enterprise
