+++ easystore-app/app/[locale]/components/molecules/confirm-register/PlanEnterprise.tsx
